Unifying handler tests in one
